# docs/swagger.yaml
definitions:
  controller.CatErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      msg:
        type: string
    type: object
  controller.CategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/dto.Category'
        type: array
    type: object
  controller.CategoryResponse:
    properties:
      category:
        $ref: '#/definitions/dto.Category'
    type: object
  controller.ErrorResponse:
    properties:
      msg:
        type: string
    type: object
  controller.ProductCountResponse:
    properties:
      product_count:
        type: integer
    type: object
  controller.SuccessResponse:
    properties:
      msg:
        type: string
    type: object
...
  dto.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      product_count:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      is_deleted:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      product_count:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
  model.CategoryInput:
    properties:
      description:
        type: string
      display_order:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
    required:
    - name
    type: object
...
  model.DashboardStats:
    properties:
      category_distribution:
        additionalProperties:
          type: integer
        type: object
      last_updated:
        type: string
      recent_reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      review_count:
        type: integer
      top_rated_products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      wishlist_count:
        type: integer
    type: object
...
  model.Product:
    properties:
      attributes:
        additionalProperties: {}
        type: object
      categories:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      cost_price:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      price:
        type: number
      sale_price:
        type: number
      sku:
        type: string
      status:
        type: string
      stock_quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      helpful_votes:
        type: integer
      id:
        type: string
      is_deleted:
        type: boolean
      is_verified_purchase:
        type: boolean
      product_id:
        type: string
      product_name:
        type: string
      rating:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Fiber go web framework based REST API boilerplate
  title: Fiber Go API
  version: "1.0"
paths:
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: Get all categories.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CategoriesResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
      summary: get all categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new category.
      parameters:
      - description: Create new category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CategoryResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: create a new category
      tags:
      - Category
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category.
      parameters:
      - description: Category ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success message
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "409":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: delete a category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get a category by ID.
      parameters:
      - description: Category ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CategoryResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
      summary: get a category by ID
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update a category.
      parameters:
      - description: Category ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Update category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CategoryResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: update a category
      tags:
      - Category
  /api/v1/categories/{id}/product-count:
    get:
      consumes:
      - application/json
      description: Get the count of products in a category.
      parameters:
      - description: Category ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: product count
          schema:
            $ref: '#/definitions/controller.ProductCountResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
      summary: get product count for a category
      tags:
      - Category
  /api/v1/categories/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get a category by slug.
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CategoryResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.CatErrorResponse'
      summary: get a category by slug
      tags:
      - Category
  /api/v1/dashboard/stats:
    get:
      consumes:
      - application/json
      description: Get dashboard statistics for current user
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard statistics
          schema:
            $ref: '#/definitions/model.DashboardStats'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: get user dashboard stats
      tags:
      - Dashboard
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"